Move scanBallot test helper to shared location
